Move log options
