+++ src/docker/functions/container/container-logs.ts
@@ -4,6 +4,7 @@ import Dockerode from "dockerode";
 import { demuxStream } from "../demux-stream";
 import { Readable } from "stream";
 import { dockerClient } from "../../docker-client";
+import { InspectResult } from "./inspect-container";
 
 export const containerLogs = async (
   container: Dockerode.Container,
@@ -17,4 +18,18 @@ export const containerLogs = async (
     log.error(`Failed to get container logs: ${err}`);
     throw err;
   }
+};
+
+export const containerRestartedLogOptions = (
+  inspectResult: InspectResult
+): Omit<Dockerode.ContainerLogsOptions, "follow" | "stdout" | "stderr"> | undefined => {
+  const { finishedAt } = inspectResult.state;
+  if (finishedAt === undefined) {
+    return undefined;
+  }
+
+  return {
+    since: finishedAt.getTime() / 1000,
+    timestamps: true,
+  };
 };
